refactor: seperate code into individual files
