+++ src/tracepoint/event.h
@@ -1,344 +1,17 @@
-#include <pthread.h>
-
-struct current_task  {
-    __u64 timestamp;
-    __u64 cgroup_id;
-    __u32 ns_id;
-    __u32 ppid;
-    __u32 pid;
-    __u32 tid;
-    __u32 uid;
-    __u32 gid;
-};
-
-struct event_t {
-    struct current_task task;
-    __s32 event_id;
-    __s64 ret;
-
-    __s32 arg_s32[6];
-    __u32 arg_u32[6];
-    __u64 arg_u64[6];
-    __u8 arg_str[256];
-
-    bool is_valid;
-    bool is_null;
-
-    __u32 ip;
-    __u16 port;
-    __u16 addr_family;
-    __u8 ipv6_addr[16];
-};
-
-struct event_key {
-    __u64 ns_id;
-    __s32 event_id;
-    char argument[256];
-};
-
-struct map_key {
-    __u32 pid;
-    __u32 tid;
-    __u32 ns_id;
-};
-
-struct getsockopt_args {
-    void *optval_ptr;
-    __u32 *optlen_ptr;
-};
-
-struct sock_addr_args {
-    void *addr_ptr;
-    __u64 *addrlen_ptr;
-};
-
-struct msg_args {
-    struct msghdr *msg_ptr;
-};
-
-struct poll_args {
-    struct pollfd *fds;
-};
-
-struct timespec_args {
-    __kernel_time64_t tv_sec;
-    long tv_nsec;
-};
-
-struct epoll_args {
-    struct epoll_event *events;
-};
-
-struct thread_ctx {
-    int thread_id;
-    struct ring_buffer *rb;  // 모든 스레드가 같은 링버퍼 공유
-    unsigned long events;
-    pthread_mutex_t *mutex;
-};
-
-struct EventEntry {
-    const char *name;
-    __u32 id;
-};
-
 #ifndef EVENT_H
 #define EVENT_H
 
-// network events
-#define SYS_ENTER_SOCKET 1
-#define SYS_EXIT_SOCKET 2
-#define SYS_ENTER_SOCKETPAIR 3
-#define SYS_EXIT_SOCKETPAIR 4
-#define SYS_ENTER_SETSOCKOPT 5
-#define SYS_EXIT_SETSOCKOPT 6
-#define SYS_ENTER_GETSOCKOPT 7
-#define SYS_EXIT_GETSOCKOPT 8
-#define SYS_ENTER_GETSOCKNAME 9
-#define SYS_EXIT_GETSOCKNAME 10
-#define SYS_ENTER_GETPEERNAME 11
-#define SYS_EXIT_GETPEERNAME 12
-#define SYS_ENTER_BIND 13
-#define SYS_EXIT_BIND 14
-#define SYS_ENTER_LISTEN 15
-#define SYS_EXIT_LISTEN 16
-#define SYS_ENTER_ACCEPT 17
-#define SYS_EXIT_ACCEPT 18
-#define SYS_ENTER_ACCEPT4 19
-#define SYS_EXIT_ACCEPT4 20
-#define SYS_ENTER_CONNECT 21
-#define SYS_EXIT_CONNECT 22
-#define SYS_ENTER_SHUTDOWN 23
-#define SYS_EXIT_SHUTDOWN 24
-#define SYS_ENTER_RECVFROM 25
-#define SYS_EXIT_RECVFROM 26
-#define SYS_ENTER_RECVMSG 27
-#define SYS_EXIT_RECVMSG 28
-#define SYS_ENTER_RECVMMSG 29
-#define SYS_EXIT_RECVMMSG 30
-#define SYS_ENTER_SENDTO 31
-#define SYS_EXIT_SENDTO 32
-#define SYS_ENTER_SENDMSG 33
-#define SYS_EXIT_SENDMSG 34
-#define SYS_ENTER_SENDMMSG 35
-#define SYS_EXIT_SENDMMSG 36
-#define SYS_ENTER_SETHOSTNAME 37
-#define SYS_EXIT_SETHOSTNAME 38
-#define SYS_ENTER_SETDOMAINNAME 39
-#define SYS_EXIT_SETDOMAINNAME 40
-#define SYS_ENTER_IOCTL 41
-#define SYS_EXIT_IOCTL 42
-#define SYS_ENTER_POLL 43
-#define SYS_EXIT_POLL 44
-#define SYS_ENTER_PPOLL 45
-#define SYS_EXIT_PPOLL 46
-#define SYS_ENTER_EPOLL_CREATE 47
-#define SYS_EXIT_EPOLL_CREATE 48
-#define SYS_ENTER_EPOLL_CREATE1 49
-#define SYS_EXIT_EPOLL_CREATE1 50
-#define SYS_ENTER_EPOLL_CTL 51
-#define SYS_EXIT_EPOLL_CTL 52
-#define SYS_ENTER_EPOLL_WAIT 53
-#define SYS_EXIT_EPOLL_WAIT 54
-#define SYS_ENTER_EPOLL_PWAIT 55
-#define SYS_EXIT_EPOLL_PWAIT 56
-#define SYS_ENTER_EPOLL_PWAIT2 57
-#define SYS_EXIT_EPOLL_PWAIT2 58
-// file events
-#define SYS_ENTER_CLOSE 101   // close_range
-#define SYS_EXIT_CLOSE 102
-#define SYS_ENTER_CREAT 103
-#define SYS_EXIT_CREAT 104
-#define SYS_ENTER_OPEN 105
-#define SYS_EXIT_OPEN 106
-#define SYS_ENTER_OPENAT 107
-#define SYS_EXIT_OPENAT 108
-#define SYS_ENTER_OPENAT2 109
-#define SYS_EXIT_OPENAT2 110
-#define SYS_ENTER_NAME_TO_HANDLE_AT 111
-#define SYS_EXIT_NAME_TO_HANDLE_AT 112
-#define SYS_ENTER_OPEN_BY_HANDLE_AT 113
-#define SYS_EXIT_OPEN_BY_HANDLE_AT 114
-#define SYS_ENTER_MEMFD_CREATE 115
-#define SYS_EXIT_MEMFD_CREATE 116
-#define SYS_ENTER_MKNOD 117
-#define SYS_EXIT_MKNOD 118
-#define SYS_ENTER_MKNODAT 119
-#define SYS_EXIT_MKNODAT 120
-#define SYS_ENTER_RENAME 121
-#define SYS_EXIT_RENAME 122
-#define SYS_ENTER_RENAMEAT 123
-#define SYS_EXIT_RENAMEAT 124
-#define SYS_ENTER_RENAMEAT2 125
-#define SYS_EXIT_RENAMEAT2 126
-#define SYS_ENTER_TRUNCATE 127
-#define SYS_EXIT_TRUNCATE 128
-#define SYS_ENTER_FTRUNCATE 129
-#define SYS_EXIT_FTRUNCATE 130
-#define SYS_ENTER_FALLOCATE 131
-#define SYS_EXIT_FALLOCATE 132
-#define SYS_ENTER_MKDIR 133
-#define SYS_EXIT_MKDIR 134
-#define SYS_ENTER_MKDIRAT 135
-#define SYS_EXIT_MKDIRAT 136
-#define SYS_ENTER_RMDIR 137
-#define SYS_EXIT_RMDIR 138
-#define SYS_ENTER_GETCWD 139
-#define SYS_EXIT_GETCWD 140
-#define SYS_ENTER_CHDIR 141
-#define SYS_EXIT_CHDIR 142
-#define SYS_ENTER_FCHDIR 143
-#define SYS_EXIT_FCHDIR 144
-#define SYS_ENTER_CHROOT 145
-#define SYS_EXIT_CHROOT 146
-#define SYS_ENTER_PIVOT_ROOT 147
-#define SYS_EXIT_PIVOT_ROOT 148
-#define SYS_ENTER_GETDENTS 149
-#define SYS_EXIT_GETDENTS 150
-#define SYS_ENTER_GETDENTS64 151
-#define SYS_EXIT_GETDENTS64 152
-#define SYS_ENTER_LINK 153   // mq_unlink
-#define SYS_EXIT_LINK 154
-#define SYS_ENTER_LINKAT 155
-#define SYS_EXIT_LINKAT 156
-#define SYS_ENTER_SYMLINK 157
-#define SYS_EXIT_SYMLINK 158
-#define SYS_ENTER_SYMLINKAT 159
-#define SYS_EXIT_SYMLINKAT 160
-#define SYS_ENTER_UNLINK 161
-#define SYS_EXIT_UNLINK 162
-#define SYS_ENTER_UNLINKAT 163
-#define SYS_EXIT_UNLINKAT 164
-#define SYS_ENTER_READLINK 165
-#define SYS_EXIT_READLINK 166
-#define SYS_ENTER_READLINKAT 167
-#define SYS_EXIT_READLINKAT 168
-#define SYS_ENTER_UMASK 169
-#define SYS_EXIT_UMASK 170
-#define SYS_ENTER_NEWSTAT 171   // cachestat(not using), statmount(not using), ustat(not using)
-#define SYS_EXIT_NEWSTAT 172
-#define SYS_ENTER_NEWLSTAT 173
-#define SYS_EXIT_NEWLSTAT 174
-#define SYS_ENTER_NEWFSTAT 175
-#define SYS_EXIT_NEWFSTAT 176
-#define SYS_ENTER_NEWFSTATAT 177
-#define SYS_EXIT_NEWFSTATAT 178
-#define SYS_ENTER_STATX 179
-#define SYS_EXIT_STATX 180
-#define SYS_ENTER_STATFS 181
-#define SYS_EXIT_STATFS 182
-#define SYS_ENTER_FSTATFS 183
-#define SYS_EXIT_FSTATFS 184
-#define SYS_ENTER_CHMOD 185
-#define SYS_EXIT_CHMOD 186
-#define SYS_ENTER_FCHMOD 187
-#define SYS_EXIT_FCHMOD 188
-#define SYS_ENTER_FCHMODAT 189  // fchmodat2(not implemented)
-#define SYS_EXIT_FCHMODAT 190
-#define SYS_ENTER_CHOWN 191
-#define SYS_EXIT_CHOWN 192
-#define SYS_ENTER_LCHOWN 193
-#define SYS_EXIT_LCHOWN 194
-#define SYS_ENTER_FCHOWN 195
-#define SYS_EXIT_FCHOWN 196
-#define SYS_ENTER_FCHOWNAT 197
-#define SYS_EXIT_FCHOWNAT 198
-#define SYS_ENTER_UTIME 199
-#define SYS_EXIT_UTIME 200
-#define SYS_ENTER_UTIMES 201
-#define SYS_EXIT_UTIMES 202
-#define SYS_ENTER_FUTIMESAT 203
-#define SYS_EXIT_FUTIMESAT 204
-#define SYS_ENTER_UTIMENSAT 205
-#define SYS_EXIT_UTIMENSAT 206
-#define SYS_ENTER_ACCESS 207
-#define SYS_EXIT_ACCESS 208
-#define SYS_ENTER_FACCESSAT 209
-#define SYS_EXIT_FACCESSAT 210  // faccessat2(not implemented)
-#define SYS_ENTER_SETXATTR 211
-#define SYS_EXIT_SETXATTR 212
-#define SYS_ENTER_LSETXATTR 213
-#define SYS_EXIT_LSETXATTR 214
-#define SYS_ENTER_FSETXATTR 215
-#define SYS_EXIT_FSETXATTR 216
-#define SYS_ENTER_GETXATTR 217
-#define SYS_EXIT_GETXATTR 218
-#define SYS_ENTER_LGETXATTR 219
-#define SYS_EXIT_LGETXATTR 220
-#define SYS_ENTER_FGETXATTR 221
-#define SYS_EXIT_FGETXATTR 222
-#define SYS_ENTER_LISTXATTR 223
-#define SYS_EXIT_LISTXATTR 224
-#define SYS_ENTER_LLISTXATTR 225
-#define SYS_EXIT_LLISTXATTR 226
-#define SYS_ENTER_FLISTXATTR 227
-#define SYS_EXIT_FLISTXATTR 228
-#define SYS_ENTER_REMOVEXATTR 229
-#define SYS_EXIT_REMOVEXATTR 230
-#define SYS_ENTER_LREMOVEXATTR 231
-#define SYS_EXIT_LREMOVEXATTR 232
-#define SYS_ENTER_FREMOVEXATTR 233
-#define SYS_EXIT_FREMOVEXATTR 234
-#define SYS_ENTER_FCNTL 235
-#define SYS_EXIT_FCNTL 236
-#define SYS_ENTER_DUP 237
-#define SYS_EXIT_DUP 238
-#define SYS_ENTER_DUP2 239
-#define SYS_EXIT_DUP2 240
-#define SYS_ENTER_DUP3 241
-#define SYS_EXIT_DUP3 242
-#define SYS_ENTER_FLOCK 243
-#define SYS_EXIT_FLOCK 244
-#define SYS_ENTER_READ 245
-#define SYS_EXIT_READ 246
-#define SYS_ENTER_PREAD64 247   // pread(name changed to pread64)
-#define SYS_EXIT_PREAD64 248
-#define SYS_ENTER_READV 249
-#define SYS_EXIT_READV 250
-#define SYS_ENTER_PREADV 251
-#define SYS_EXIT_PREADV 252
-#define SYS_ENTER_PREADV2 253
-#define SYS_EXIT_PREADV2 254
-#define SYS_ENTER_WRITE 255
-#define SYS_EXIT_WRITE 256
-#define SYS_ENTER_PWRITE64 257  // pwrite(name changed to pwrite64)
-#define SYS_EXIT_PWRITE64 258
-#define SYS_ENTER_WRITEV 259
-#define SYS_EXIT_WRITEV 260
-#define SYS_ENTER_PWRITEV 261
-#define SYS_EXIT_PWRITEV 262
-#define SYS_ENTER_PWRITEV2 263
-#define SYS_EXIT_PWRITEV2 264
-#define SYS_ENTER_LSEEK 265
-#define SYS_EXIT_LSEEK 266
-#define SYS_ENTER_SENDFILE64 267    // sendfile(name changed to sendfile64)
-#define SYS_EXIT_SENDFILE64 268
-#define SYS_ENTER_INOTIFY_INIT 269
-#define SYS_EXIT_INOTIFY_INIT 270
-#define SYS_ENTER_INOTIFY_INIT1 271
-#define SYS_EXIT_INOTIFY_INIT1 272
-#define SYS_ENTER_INOTIFY_ADD_WATCH 273
-#define SYS_EXIT_INOTIFY_ADD_WATCH 274
-#define SYS_ENTER_INOTIFY_RM_WATCH 275
-#define SYS_EXIT_INOTIFY_RM_WATCH 276
-#define SYS_ENTER_FANOTIFY_INIT 277
-#define SYS_EXIT_FANOTIFY_INIT 278
-#define SYS_ENTER_FANOTIFY_MARK 279
-#define SYS_EXIT_FANOTIFY_MARK 280
-#define SYS_ENTER_MOUNT 281
-#define SYS_EXIT_MOUNT 282
-#define SYS_ENTER_UMOUNT 283    // umount2(not using)
-#define SYS_EXIT_UMOUNT 284
-#define SYS_ENTER_MOVE_MOUNT 285
-#define SYS_EXIT_MOVE_MOUNT 286
-// Synchronized I/O, Asynchronous I/O, Miscellaneous, Multiplexed I/O events is not included -> not essential
-// process events
-#define SYS_ENTER_MMAP 909
-#define SYS_EXIT_MMAP 910
-#define SYS_ENTER_MUNMAP 911
-#define SYS_EXIT_MUNMAP 912
-#define SYS_ENTER_MPROTECT 913
-#define SYS_EXIT_MPROTECT 914
-#define SYS_ENTER_PKEY_MPROTECT 915
-#define SYS_EXIT_PKEY_MPROTECT 916
+#include <stdio.h>
+#include <string.h>
+#include <stdbool.h>
+#include <sys/sysinfo.h>
+#include "syscalls.h"
+#include "structs.h"
+
+extern struct EventEntry event_table[];
+
+__u32 hash(const char* str);
+void init_event_table(void);
+__u32 get_event_id(const char* event_str);
 
 #endif
